backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1


# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40

# OpenAI
openai>=1.10.0

# Web Search
tavily-python>=0.3.0

# Observability
langfuse>=2.0.0

# SSE
sse-starlette>=1.8.2

# Utilities
httpx>=0.26.0
python-dotenv>=1.0.0

# Code validation
ruff>=0.1.14

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0
